+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 ## 0.0.4
 
 * Streaming api for meeting status
+## 0.0.5
+
+* Enhancements
+
+    * Additional flags:   
+  ```
+  String noDisconnectAudio; 
+  String noAudio;
+  ```
+    * Upgrade Zoom to support GCM.
+    * Drop the android min sdk to 21